minSdkVersion 16, use supported CMake NDK build chain
The gradle-experimental plugin is already unsupported, see
http://tools.android.com/tech-docs/new-build-system/gradle-experimental
and it was breaking release builds in a non-obvious way relating to Proguard.

This also massively simplifies building the executable; CMake can just do that.

We also require API 14 (4.0) now and it would be complicated to keep supporting 14-15 (non-PIE) in CMake as well as APIs 21+ (5.0+) that **require** PIE, so I've bumped minSdkVersion to 16 (4.1) so we can just leave PIE always on.